Attempting to change url appearance
So rn i just Ctrl C and Ctrl V this section of code from index.html and added it to "WhoamI.html" to see if this will affect the name that appears in the URL bar. Otherwise im just experimenting and seeing what happenes. Will revert this change if it breaks the website.